feat: Add Suspense wrapper and reorganize folder structure
